# swagger.yaml
openapi: 3.0.0
info:
  title: Wallet API
  version: 1.0.0
  description: Wallet / iGaming wallet API (users, deposits, withdrawals, bets, transaction history)
servers:
  - url: http://localhost:3000
    description: Local dev
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
          example: "Amount must be a positive number"
    User:
      type: object
      properties:
        id:
          type: string
          example: "64a1f8b2a1b2c3d4e5f67890"
        username:
          type: string
          example: "alice"
        balance:
          type: number
          example: 150
        kycApproved:
          type: boolean
          example: false
        createdAt:
          type: string
          format: date-time
    AuthResponse:
      type: object
      properties:
        token:
          type: string
          example: "eyJhbGciOiJI..."
        user:
          $ref: '#/components/schemas/User'
...
paths:
  /api/auth/register:
    post:
      tags: [Auth]
      summary: Register (public)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username, password]
              properties:
                username: { type: string }
                password: { type: string }
      responses:
        '201':
          description: Created — returns token + user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: Username taken
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/auth/login:
    post:
      tags: [Auth]
      summary: Login (public)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username, password]
              properties:
                username: { type: string }
                password: { type: string }
      responses:
        '200':
          description: Authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...